requirements.txt
sentence-transformers==2.2.2
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-pinecone==0.1.3
langchain-google-genai==1.0.10
flask
pypdf
python-dotenv
pinecone[grpc]